+++ includes/class-ldd-ordering.php
@@ -79,9 +79,10 @@ class LDD_Ordering extends Aihrus_Common {
 
 	public static function actions() {
 		// fixme add_action( 'admin_menu', array( __CLASS__, 'admin_menu' ) );
-		add_action( 'edd_checkout_error_checks', array( __CLASS__, 'edd_checkout_error_checks' ), 10, 2 );
 		add_action( 'admin_enqueue_scripts', array( __CLASS__, 'scripts' ) );
 		add_action( 'admin_init', array( __CLASS__, 'admin_init' ) );
+		add_action( 'edd_after_checkout_cart', array( __CLASS__, 'edd_after_checkout_cart' ) );
+		add_action( 'edd_checkout_error_checks', array( __CLASS__, 'edd_checkout_error_checks' ), 10, 2 );
 		add_action( 'edd_post_add_to_cart', array( __CLASS__, 'edd_post_add_to_cart' ), 10, 2 );
 		add_action( 'edd_post_remove_from_cart', array( __CLASS__, 'edd_post_remove_from_cart' ), 10, 2 );
 		add_action( 'edd_update_payment_status', array( __CLASS__, 'edd_update_payment_status' ), 10, 3 );
@@ -635,9 +636,13 @@ class LDD_Ordering extends Aihrus_Common {
 		}
 
 		$delivery_intersect = array_intersect( $delivery_options, $ordered_items );
-		if ( 1 < count( $delivery_intersect ) ) {
+		$delivery_items     = count( $delivery_intersect );
+		if ( 1 < $delivery_items ) {
 			$text = esc_html__( 'Please remove a delivery option. Only one is allowed per order.' );
 			edd_set_error( 'excess_delivery_item', $text );
+		} elseif ( empty( $delivery_items ) ) {
+			$text = __( 'Please <a href="/#services">choose a delivery option</a>. One is required per order.' );
+			edd_set_error( 'no_delivery_item', $text );
 		}
 
 		if ( ! empty( $post['cfm_files']['court_filings'] ) ) {
@@ -658,6 +663,25 @@ class LDD_Ordering extends Aihrus_Common {
 
 			$text = esc_html__( 'No documents uploaded for filing' );
 			edd_set_error( 'no_documents', $text );
+		}
+	}
+
+
+	public static function edd_after_checkout_cart() {
+		$cart   = edd_get_cart_contents();
+		$fee_id = ldd_get_option( 'filing_fee_id' );
+		$show   = true;
+		foreach ( $cart as $key => $item ) {
+			if ( $fee_id == $item['id'] ) {
+				$show = false;
+			}
+		}
+
+		if ( $show ) {
+			echo '<div class="ldd-ordering">';
+			echo '<h3>Have Filings Fees? Add Them Here!</h3>';
+			echo do_shortcode( '[purchase_link id="241"]' );
+			echo '</div>';
 		}
 	}
 }